sample refactor
